Switch to JSPI + offscreen canvas + main in worker

+++ install_emsdk.sh
@@ -4,6 +4,19 @@ BASE_DIR="$(dirname -- "$(readlink -f -- "$0")")"
 
 cd "$BASE_DIR"
 
+# Node check
+NODE_FULL_PATH="$(which node)"
+INSTALLED_NODE_VERSION=`"$NODE_FULL_PATH" --version || true`
+if [[ "$INSTALLED_NODE_VERSION" != v24.* ]]; then
+    set +eux
+    echo
+    echo "Configure will fail unless Node.js v24.18.0 (or newer) is installed and in the PATH"
+    echo "Visit https://nodejs.org/en/download for install instructions"
+    echo "A newer version may also work. Update install_emsdk.sh as needed."
+    echo
+    exit 1
+fi
+
 rm -rf emsdk
 git clone https://github.com/emscripten-core/emsdk.git
 
@@ -14,3 +27,11 @@ popd
 
 ./apply_patches.sh emsdk
 
+pushd emsdk
+sed -i "s|^NODE_JS = .*|NODE_JS = '$NODE_FULL_PATH'|" .emscripten
+popd
+
+# Rebuild library to incorporate "emsdk_align.patch"
+pushd emsdk
+upstream/emscripten/embuilder build libhtml5 --force
+popd
